Feature bm security

+++ src/main/java/ru/skypro/homework/service/impl/CommentServiceImpl.java
@@ -2,6 +2,7 @@ package ru.skypro.homework.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import ru.skypro.homework.dto.Comment;
 import ru.skypro.homework.dto.ResponseWrapperComment;
@@ -28,6 +29,7 @@ public class CommentServiceImpl implements CommentService {
     private final CommentMapper commentMapper;
     private final AdsRepository adsRepository;
     private final UserRepository userRepository;
+    private final UserValidatePermission validatePermission;
     private final DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("dd-MM-yyyy HH:mm:ss");
 
 
@@ -53,11 +55,10 @@ public class CommentServiceImpl implements CommentService {
         UserEntity findUser = userRepository.findUserEntityByEmail(authentication.getName())
                 .orElseThrow(() -> new UserNotRegisterException(authentication.getName()));
 
-        if (findComment.getUser().getId().equals(findUser.getId())) {
+        if(validatePermission.isAdmin(findUser) || validatePermission.isCommentOwner(findUser, findComment)) {
             findComment.setText(comment.getText());
             findComment.setCreatedAt(LocalDateTime.now());
             return commentMapper.modelToDto(commentRepository.save(findComment));
-
         } else {
             throw new UserForbiddenException(findUser.getId());
         }
@@ -69,9 +70,15 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public void deleteComment(Integer adId, Integer commentId) {
-        findComment(adId, commentId);
-        commentRepository.deleteById(commentId);
+    public void deleteComment(Integer adId, Integer commentId, Authentication authentication) {
+        CommentEntity comment = findComment(adId, commentId);
+        UserEntity user = userRepository.findUserEntityByEmail(authentication.getName())
+                .orElseThrow(() -> new UsernameNotFoundException("User not found: " + authentication.getName()));
+        if (validatePermission.isAdmin(user) || validatePermission.isCommentOwner(user, comment)) {
+            commentRepository.deleteById(commentId);
+            return;
+        }
+        throw new UserForbiddenException(user.getId());
     }
 
     @Override
